fix: fix `configure-test-app` crashing on 0.75
