+++ package.json
@@ -53,6 +53,7 @@
     "scripts/parseargs.mjs",
     "scripts/schema.mjs",
     "scripts/template.mjs",
+    "scripts/utils/npm.mjs",
     "test-app.gradle",
     "test_app.rb",
     "visionos",
