+++ scripts/set-react-version.mjs
@@ -8,14 +8,13 @@ import { promises as fs } from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 import {
-  fetchPackageMetadata,
   isMain,
-  npmRegistryBaseURL,
   readJSONFile,
   readTextFile,
   toVersionNumber,
   v,
 } from "./helpers.js";
+import { fetchPackageMetadata, npmRegistryBaseURL } from "./utils/npm.mjs";
 
 /**
  * @typedef {import("./types.js").Manifest} Manifest
